dividing routing
