+++ src/main/resources/application.conf
@@ -7,7 +7,7 @@ akka.persistence.journal.plugin = "akka-contrib-mongodb-persistence-journal"
 akka.persistence.snapshot-store.plugin = "akka-contrib-mongodb-persistence-snapshot"
 
 akka.contrib.persistence.mongodb.mongo.driver = "akka.contrib.persistence.mongodb.RxMongoPersistenceExtension"
-akka.contrib.persistence.mongodb.mongo.mongouri = "mongodb://192.168.59.103:27017/replicated-database"
+akka.contrib.persistence.mongodb.mongo.mongouri = "mongodb://localhost:27017/replicated-database"
 
 akka.persistence.journal.leveldb.native = off
 akka.persistence.journal.leveldb-shared.store.native = off
